add unit test for DefaultMessageStore
